+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   - `--scd` argument enabling scene change detection.
   - `--svt ARG` for additional args, _e.g. `--svt mbr=2000 --svt film-grain=30`_.
 * Rename _vmaf_ command argument `--reference` (was `--original`).
+* Use 128k bitrate as a default for libopus audio.
+* Remove `--aq`.
 * Add optional `--vmaf-options` argument to _vmaf, sample-encode, crf-search, auto-encode_ commands.
 * Fail fast if ffmpeg cut samples are empty (< 1K).
 * Handle input durations lower than the sample duration by using the whole input as a single sample.